refactor: util method 추가 분리 - #190
