+++ src/contents/contents.service.ts
@@ -5,9 +5,6 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { EntityManager } from 'typeorm';
-import * as cheerio from 'cheerio';
-import axios from 'axios';
-import * as fs from 'fs';
 
 import {
   AddCategoryBodyDto,
@@ -46,6 +43,7 @@ import { UserRepository } from '../users/repository/user.repository';
 import { ContentRepository } from './repository/content.repository';
 import { CategoryUtil } from './util/category.util';
 import { CategoryRepository } from './repository/category.repository';
+import { ContentUtil } from './util/content.util';
 
 @Injectable()
 export class ContentsService {
@@ -55,6 +53,7 @@ export class ContentsService {
     private readonly summaryService: SummaryService,
     private readonly categoryRepository: CategoryRepository,
     private readonly categoryUtil: CategoryUtil,
+    private readonly contentUtil: ContentUtil,
   ) {}
 
   async addContent(
@@ -83,7 +82,7 @@ export class ContentsService {
         siteName,
         description,
         coverImg,
-      } = await this.getLinkInfo(link);
+      } = await this.contentUtil.getLinkInfo(link);
       title = title ? title : linkTitle;
 
       let category: Category | null = null;
@@ -152,7 +151,7 @@ export class ContentsService {
         }
         for (const link of contentLinks) {
           const { title, description, coverImg, siteName } =
-            await this.getLinkInfo(link);
+            await this.contentUtil.getLinkInfo(link);
 
           if (category) {
             await this.categoryUtil.checkContentDuplicateAndAddCategorySaveLog(
@@ -344,59 +343,6 @@ export class ContentsService {
     } catch (e) {
       throw e;
     }
-  }
-
-  async getLinkInfo(link: string) {
-    let title: string | undefined = '';
-    let coverImg: string | undefined = '';
-    let description: string | undefined = '';
-    let siteName: string | undefined;
-
-    await axios
-      .get(link)
-      .then((res) => {
-        if (res.status !== 200) {
-          throw new BadRequestException('잘못된 링크입니다.');
-        } else {
-          const data = res.data;
-          if (typeof data === 'string') {
-            const $ = cheerio.load(data);
-            title = $('title').text() !== '' ? $('title').text() : 'Untitled';
-            $('meta').each((i, el) => {
-              const meta = $(el);
-              if (meta.attr('property') === 'og:image') {
-                coverImg = meta.attr('content');
-              }
-              if (meta.attr('property') === 'og:description') {
-                description = meta.attr('content');
-              }
-              if (meta.attr('property') === 'og:site_name') {
-                siteName = meta.attr('content');
-              }
-            });
-          }
-        }
-      })
-      .catch((e) => {
-        // Control unreachable link
-        // if(e.message === 'Request failed with status code 403') {
-        // 403 에러가 발생하는 링크는 크롤링이 불가능한 링크이다.
-        // }
-        for (let idx = 1; idx < 3; idx++) {
-          if (link.split('/').at(-idx) !== '') {
-            title = link.split('/').at(-idx);
-            break;
-          }
-        }
-        title = title ? title : 'Untitled';
-      });
-
-    return {
-      title,
-      description,
-      coverImg,
-      siteName,
-    };
   }
 
   async loadPersonalContents(
@@ -804,7 +750,8 @@ export class CategoryService {
   ): Promise<LoadFrequentCategoriesOutput> {
     try {
       // 로그 파일 내의 기록을 불러온다.
-      const recentCategoryList: RecentCategoryList[] = this.loadLogs(user.id);
+      const recentCategoryList: RecentCategoryList[] =
+        this.categoryUtil.loadLogs(user.id);
 
       // 캐시 내의 카테고리 리스트를 최신 순으로 정렬하고, 동시에 저장된 횟수를 추가한다.
 
@@ -822,11 +769,12 @@ export class CategoryService {
 
         // 10개의 로그를 확인한다.
         i += 10;
-        recentCategoriesWithSaveCount = this.makeCategoryListWithSaveCount(
-          recentCategoryList,
-          recentCategoriesWithSaveCount,
-          i,
-        );
+        recentCategoriesWithSaveCount =
+          this.categoryUtil.makeCategoryListWithSaveCount(
+            recentCategoryList,
+            recentCategoriesWithSaveCount,
+            i,
+          );
         // 10개의 로그를 확인했으므로 남은 로그 수를 10개 감소시킨다.
         remainLogCount -= 10;
 
@@ -899,62 +847,5 @@ export class CategoryService {
     } catch (e) {
       throw e;
     }
-  }
-
-  /**
-   * 파일에서 로그를 불러오는 함수
-   * @param id
-   * @returns RecentCategoryList[]
-   */
-  loadLogs(id: number): RecentCategoryList[] {
-    const logList: string[] = fs
-      .readFileSync(`${__dirname}/../../user_logs/${id}.txt`)
-      .toString()
-      .split('\n');
-    logList.pop(); // 마지막 줄은 빈 줄이므로 제거
-
-    // logList를 RecentCategoryList[]로 변환
-    const recentCategoryList: RecentCategoryList[] = logList.map((str) => {
-      const categoryId = +str.split('"categoryId": ')[1].split(',')[0];
-      const savedAt = +str.split('"savedAt": ')[1].split('}')[0];
-      return {
-        categoryId,
-        savedAt,
-      };
-    });
-
-    // 최신 순으로 정렬 후 반환
-    return recentCategoryList.reverse();
-  }
-
-  /**
-   * 불러온 로그를 바탕으로 카테고리당 저장된 카운트와 함께 배열을 만드는 함수(매번 10개씩 조회한다.)
-   * @param recentCategoryList
-   * @param recentCategoriesWithSaveCount
-   * @param till
-   * @returns
-   */
-  makeCategoryListWithSaveCount(
-    recentCategoryList: RecentCategoryList[],
-    recentCategoriesWithSaveCount: RecentCategoryListWithSaveCount[],
-    till: number,
-  ): RecentCategoryListWithSaveCount[] {
-    const start: number = till - 10;
-    const end: number = till;
-    for (let i = start; i < end && i < recentCategoryList.length; i++) {
-      const inNewList = recentCategoriesWithSaveCount.find(
-        (category) => category.categoryId === recentCategoryList[i].categoryId,
-      );
-      if (inNewList) {
-        inNewList.saveCount++;
-      } else {
-        recentCategoriesWithSaveCount.push({
-          ...recentCategoryList[i],
-          saveCount: 1,
-        });
-      }
-    }
-
-    return recentCategoriesWithSaveCount;
   }
 }
